+++ tomcat/src/main/java/nextstep/jwp/vo/FileName.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package nextstep.jwp.vo;
 
 public class FileName {
