+++ tomcat/src/main/java/nextstep/jwp/vo/HttpCookie.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package nextstep.jwp.vo;
 
 import java.util.Map;
